# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ParseStatus": {
        "enum": [
          "pending",
          "parsing",
          "complete",
          "failed"
        ],
        "title": "ParseStatus",
        "type": "string"
      },
...
      "RepositoryOut": {
        "properties": {
          "commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Sha"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Branch"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "file_count": {
            "title": "File Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parsed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsed At"
          },
          "status": {
            "$ref": "#/components/schemas/ParseStatus"
          },
          "symbol_count": {
            "title": "Symbol Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "id",
          "name",
          "url",
          "default_branch",
          "commit_sha",
          "status",
          "error_message",
          "file_count",
          "symbol_count",
          "created_at",
          "updated_at",
          "parsed_at"
        ],
        "title": "RepositoryOut",
        "type": "object"
      },
...
      "SearchHitOut": {
        "properties": {
          "citation": {
            "title": "Citation",
            "type": "string"
          },
          "hops": {
            "description": "CALLS hops from the seed, 0 for a direct hit",
            "title": "Hops",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "line_end": {
            "title": "Line End",
            "type": "integer"
          },
          "line_start": {
            "title": "Line Start",
            "type": "integer"
          },
          "module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "qualname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualname"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source": {
            "description": "vector | graph",
            "title": "Source",
            "type": "string"
          },
          "symbol_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Neo4j node key, if a symbol",
            "title": "Symbol Key"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "via": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "the seed this was expanded from",
            "title": "Via"
          }
        },
        "required": [
          "score",
          "path",
          "kind",
          "line_start",
          "line_end",
          "text",
          "citation",
          "source",
          "hops"
        ],
        "title": "SearchHitOut",
        "type": "object"
      },
      "SearchResponse": {
        "properties": {
          "from_graph": {
            "description": "hits similarity alone would not have found",
            "title": "From Graph",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchHitOut"
            },
            "title": "Items",
            "type": "array"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "query",
          "mode",
          "items",
          "total",
          "from_graph"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "AI Engineering Intelligence Platform",
    "title": "ARGUS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/repos/{repository_id}": {
      "delete": {
        "operationId": "delete_repository_repos__repository_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Repository",
        "tags": [
          "repositories"
        ]
      },
      "get": {
        "operationId": "get_repository_repos__repository_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Repository",
        "tags": [
          "repositories"
        ]
      }
    },
...
    "/repos/{repository_id}/search": {
      "get": {
        "operationId": "semantic_search_repos__repository_id__search_get",
        "parameters": [
          {
            "in": "path",
            "name": "repository_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Repository Id",
              "type": "string"
            }
          },
          {
            "description": "A question about the codebase",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "A question about the codebase",
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "top_k",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 50,
              "minimum": 1,
              "title": "Top K",
              "type": "integer"
            }
          },
          {
            "description": "function | method | class | module",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "function | method | class | module",
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "hybrid",
              "pattern": "^(vector|hybrid)$",
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "description": "CALLS hops, hybrid only",
            "in": "query",
            "name": "depth",
            "required": false,
            "schema": {
              "default": 1,
              "description": "CALLS hops, hybrid only",
              "maximum": 2,
              "minimum": 1,
              "title": "Depth",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Semantic Search",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}